Tip: Don't fix merge conflict with Github.app, they're ugly

+++ core/app/assets/javascripts/backbone/views/fact_view.js
@@ -10,7 +10,7 @@ window.FactView = Backbone.View.extend({
     "click a.remove": "removeFactFromChannel",
     "click li.destroy": "destroyFact",
 	  "click .controls .supporting, .controls .weakening": "toggleEvidence",
-    "click .controls .supporting a.add-button,    .controls .weakening a.add-button": "showAddRelation",
+    "click .controls .supporting a.add-button,    .controls .weakening a.add-button, .no-evidence-message .do-add": "showAddRelation",
     "click .controls .supporting a.cancel-button, .controls .weakening a.cancel-button": "hideAddRelation"
   },
 
@@ -51,6 +51,7 @@ window.FactView = Backbone.View.extend({
   },
 
   removeFactFromChannel: function() {
+    if(!confirm("Do you want to remove this Factlink from your channel? It will still be available in your `Created` channel.")) return false;
     this.model.destroy({
       error: function() {
         alert("Error while removing Factlink from Channel" );
@@ -60,6 +61,7 @@ window.FactView = Backbone.View.extend({
   },
 
   destroyFact: function() {
+    if(!confirm("Are you sure you want to delete the Factlink you have created?")) return false;
     this.model.destroy({
       error: function() {
         alert("Error while removing the Factlink" );
@@ -158,7 +160,7 @@ window.FactView = Backbone.View.extend({
 
   showAddRelation: function(e) {
     e.stopPropagation();
-
+    $('.do-add').hide();
     $(e.target).toggleClass('add-button cancel-button');
 
     if ( this._currentVisibleDropdown === "supporting" ) {
@@ -170,7 +172,7 @@ window.FactView = Backbone.View.extend({
 
   hideAddRelation: function(e) {
     e.stopPropagation();
-
+$('.do-add').show();
     $(e.target).toggleClass('add-button cancel-button');
 
     if ( this._currentVisibleDropdown === "supporting" ) {
@@ -185,7 +187,7 @@ window.FactView = Backbone.View.extend({
     $(this.model.get("last_active_users")).each(function()  {
       var el = $("li.user[data-user-id="+ this.graph_id + "]", self.el);
       var model = new User(this);
-      var view = new UserPassportView({model: model, el: el}).render();
+      var view = new UserPassportView({model: model, el: el});
     });
   }
 });
